+++ app/[team]/[[...section]]/layout.tsx
@@ -7,7 +7,11 @@ export default function SectionLayout({
 }) {
   return (
     <>
-      {nav}
+      <nav className="bg-white shadow">
+        <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
+          {nav}
+        </div>
+      </nav>
       {children}
     </>
   );
